migrate world database to EF Core * restructure and cleanup
