+++ src/Rasa.DBL/Context/Auth/MySqlAuthContext.cs
@@ -1,13 +1,10 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Options;
+﻿using Microsoft.Extensions.Options;
 
 namespace Rasa.Context.Auth
 {
     using Configuration;
     using Configuration.ContextSetup;
-    using Services;
     using Services.DbContext;
-    using Structures;
 
     public class MySqlAuthContext : AuthContext
     {
