+++ src/Rasa.DBL/Context/Auth/SqliteAuthContext.cs
@@ -4,7 +4,6 @@ namespace Rasa.Context.Auth
 {
     using Configuration;
     using Configuration.ContextSetup;
-    using Services;
     using Services.DbContext;
 
     public class SqliteAuthContext : AuthContext
