+++ src/Rasa.DBL/Repositories/Auth/Account/AccountLockedException.cs
@@ -3,6 +3,7 @@
 namespace Rasa.Repositories.Auth.Account
 {
     using Structures;
+    using Structures.Auth;
 
     public class AccountLockedException : Exception
     {
