+++ src/Rasa.DBL/Repositories/Auth/Account/AuthAccountRepository.cs
@@ -11,6 +11,7 @@ namespace Rasa.Repositories.Auth.Account
     using Context.Auth;
     using Services.Random;
     using Structures;
+    using Structures.Auth;
 
     public class AuthAccountRepository : IAuthAccountRepository
     {
