+++ src/Rasa.DBL/Repositories/Auth/Account/PasswordCheckFailedException.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using Structures;
+    using Structures.Auth;
 
     public class PasswordCheckFailedException : Exception
     {
