+++ src/Rasa.DBL/Structures/Char/ClanEntry.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Rasa.Structures
+namespace Rasa.Structures.Char
 {
     using Interfaces;
