+++ src/Rasa.DBL/Structures/Char/GameAccountEntry.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Rasa.Structures
+namespace Rasa.Structures.Char
 {
     using Interfaces;
     using JetBrains.Annotations;
